[ADD] 유틸리티 함수 분리 및 정의
